Moved ghost diagnostics into a dedicated notification channel
